refactor MCP and servers to support sse and idalib
