+++ src/ida_pro_mcp/ida_tools.py
@@ -3,12 +3,9 @@ import sys
 if sys.version_info < (3, 11):
     raise RuntimeError("Python 3.11 or higher is required for the MCP plugin")
 import re
-import json
 import struct
-import threading
-import http.server
-from urllib.parse import urlparse
 from typing import Any, Callable, get_type_hints, TypedDict, Optional, Annotated, TypeVar, Generic
+
 
 class JSONRPCError(Exception):
     def __init__(self, code: int, message: str, data: Any = None):
@@ -75,157 +72,6 @@ def jsonrpc(func: Callable) -> Callable:
     global rpc_registry
     return rpc_registry.register(func)
 
-class JSONRPCRequestHandler(http.server.BaseHTTPRequestHandler):
-    def send_jsonrpc_error(self, code: int, message: str, id: Any = None):
-        response = {
-            "jsonrpc": "2.0",
-            "error": {
-                "code": code,
-                "message": message
-            }
-        }
-        if id is not None:
-            response["id"] = id
-        response_body = json.dumps(response).encode("utf-8")
-        self.send_response(200)
-        self.send_header("Content-Type", "application/json")
-        self.send_header("Content-Length", len(response_body))
-        self.end_headers()
-        self.wfile.write(response_body)
-
-    def do_POST(self):
-        global rpc_registry
-
-        parsed_path = urlparse(self.path)
-        if parsed_path.path != "/mcp":
-            self.send_jsonrpc_error(-32098, "Invalid endpoint", None)
-            return
-
-        content_length = int(self.headers.get("Content-Length", 0))
-        if content_length == 0:
-            self.send_jsonrpc_error(-32700, "Parse error: missing request body", None)
-            return
-
-        request_body = self.rfile.read(content_length)
-        try:
-            request = json.loads(request_body)
-        except json.JSONDecodeError:
-            self.send_jsonrpc_error(-32700, "Parse error: invalid JSON", None)
-            return
-
-        # Prepare the response
-        response = {
-            "jsonrpc": "2.0"
-        }
-        if request.get("id") is not None:
-            response["id"] = request.get("id")
-
-        try:
-            # Basic JSON-RPC validation
-            if not isinstance(request, dict):
-                raise JSONRPCError(-32600, "Invalid Request")
-            if request.get("jsonrpc") != "2.0":
-                raise JSONRPCError(-32600, "Invalid JSON-RPC version")
-            if "method" not in request:
-                raise JSONRPCError(-32600, "Method not specified")
-
-            # Dispatch the method
-            result = rpc_registry.dispatch(request["method"], request.get("params", []))
-            response["result"] = result
-
-        except JSONRPCError as e:
-            response["error"] = {
-                "code": e.code,
-                "message": e.message
-            }
-            if e.data is not None:
-                response["error"]["data"] = e.data
-        except IDAError as e:
-            response["error"] = {
-                "code": -32000,
-                "message": e.message,
-            }
-        except Exception as e:
-            traceback.print_exc()
-            response["error"] = {
-                "code": -32603,
-                "message": "Internal error (please report a bug)",
-                "data": traceback.format_exc(),
-            }
-
-        try:
-            response_body = json.dumps(response).encode("utf-8")
-        except Exception as e:
-            traceback.print_exc()
-            response_body = json.dumps({
-                "error": {
-                    "code": -32603,
-                    "message": "Internal error (please report a bug)",
-                    "data": traceback.format_exc(),
-                }
-            }).encode("utf-8")
-
-        self.send_response(200)
-        self.send_header("Content-Type", "application/json")
-        self.send_header("Content-Length", len(response_body))
-        self.end_headers()
-        self.wfile.write(response_body)
-
-    def log_message(self, format, *args):
-        # Suppress logging
-        pass
-
-class MCPHTTPServer(http.server.HTTPServer):
-    allow_reuse_address = False
-
-class Server:
-    HOST = "localhost"
-    PORT = 13337
-
-    def __init__(self):
-        self.server = None
-        self.server_thread = None
-        self.running = False
-
-    def start(self):
-        if self.running:
-            print("[MCP] Server is already running")
-            return
-
-        self.server_thread = threading.Thread(target=self._run_server, daemon=True)
-        self.running = True
-        self.server_thread.start()
-
-    def stop(self):
-        if not self.running:
-            return
-
-        self.running = False
-        if self.server:
-            self.server.shutdown()
-            self.server.server_close()
-        if self.server_thread:
-            self.server_thread.join()
-            self.server = None
-        print("[MCP] Server stopped")
-
-    def _run_server(self):
-        try:
-            # Create server in the thread to handle binding
-            self.server = MCPHTTPServer((Server.HOST, Server.PORT), JSONRPCRequestHandler)
-            print(f"[MCP] Server started at http://{Server.HOST}:{Server.PORT}")
-            self.server.serve_forever()
-        except OSError as e:
-            if e.errno == 98 or e.errno == 10048:  # Port already in use (Linux/Windows)
-                print("[MCP] Error: Port 13337 is already in use")
-            else:
-                print(f"[MCP] Server error: {e}")
-            self.running = False
-        except Exception as e:
-            print(f"[MCP] Server error: {e}")
-        finally:
-            self.running = False
-
 # A module that helps with writing thread safe ida code.
 # Based on:
 # https://web.archive.org/web/20160305190440/http://www.williballenthin.com/blog/2015/09/04/idapython-synchronization-decorator/
@@ -233,7 +79,7 @@ import logging
 import queue
 import traceback
 import functools
-
+  
 import ida_hexrays
 import ida_kernwin
 import ida_funcs
@@ -280,7 +126,6 @@ class IDASyncError(Exception):
 #       return idautils.Functions()
 #
 
-logger = logging.getLogger(__name__)
 
 # Enum for safety modes. Higher means safer:
 class IDASafety:
@@ -995,26 +840,3 @@ def set_local_variable_type(
         raise IDAError(f"Failed to modify local variable: {variable_name}")
     refresh_decompiler_ctext(func.start_ea)
 
-class MCP(idaapi.plugin_t):
-    flags = idaapi.PLUGIN_KEEP
-    comment = "MCP Plugin"
-    help = "MCP"
-    wanted_name = "MCP"
-    wanted_hotkey = "Ctrl-Alt-M"
-
-    def init(self):
-        self.server = Server()
-        hotkey = MCP.wanted_hotkey.replace("-", "+")
-        if sys.platform == "darwin":
-            hotkey = hotkey.replace("Alt", "Option")
-        print(f"[MCP] Plugin loaded, use Edit -> Plugins -> MCP ({hotkey}) to start the server")
-        return idaapi.PLUGIN_KEEP
-
-    def run(self, args):
-        self.server.start()
-
-    def term(self):
-        self.server.stop()
-
-def PLUGIN_ENTRY():
-    return MCP()
